# Backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/working-day/check-in": {
      "post": {
        "operationId": "WorkingDayController_checkIn_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckInDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Check In",
        "tags": [
          "WorkingDay"
        ]
      }
    },
    "/v1/working-day/check-out": {
      "post": {
        "operationId": "WorkingDayController_checkOut_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckOutDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Check Out",
        "tags": [
          "WorkingDay"
        ]
      }
    },
    "/v1/working-day/history": {
      "get": {
        "operationId": "WorkingDayController_getHistory_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Limit",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "SortBy",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "SortOrder",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "StartDate",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "EndDate",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Status",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "salesman_id",
            "required": false,
            "in": "query",
            "description": "Filter by salesman id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/WorkingDay"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get History",
        "tags": [
          "WorkingDay"
        ]
      }
    },
...
  },
  "info": {
    "title": "Vikas Inventory RN Platform API",
    "description": "Enterprise Backend API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PaginatedMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "total": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          },
          "hasNextPage": {
            "type": "boolean"
          },
          "hasPreviousPage": {
            "type": "boolean"
          }
        },
        "required": [
          "page",
          "limit",
          "total",
          "totalPages",
          "hasNextPage",
          "hasPreviousPage"
        ]
      },
      "PaginatedDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginatedMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
      "CheckInDto": {
        "type": "object",
        "properties": {
          "latitude": {
            "type": "number",
            "description": "Latitude"
          },
          "longitude": {
            "type": "number",
            "description": "Longitude"
          },
          "device_id": {
            "type": "string",
            "description": "Device id"
          },
          "idempotency_key": {
            "type": "string",
            "description": "Idempotency key"
          }
        },
        "required": [
          "latitude",
          "longitude"
        ]
      },
      "CheckOutDto": {
        "type": "object",
        "properties": {
          "latitude": {
            "type": "number",
            "description": "Latitude"
          },
          "longitude": {
            "type": "number",
            "description": "Longitude"
          },
          "device_id": {
            "type": "string",
            "description": "Device id"
          },
          "idempotency_key": {
            "type": "string",
            "description": "Idempotency key"
          }
        },
        "required": [
          "latitude",
          "longitude"
        ]
      },
      "WorkingDay": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Id"
          },
          "salesman_id": {
            "type": "string",
            "description": "Salesman id"
          },
          "distributor_id": {
            "type": "string",
            "description": "Distributor id"
          },
          "check_in_at": {
            "format": "date-time",
            "type": "string",
            "description": "Check in at"
          },
          "check_out_at": {
            "format": "date-time",
            "type": "string",
            "description": "Check out at"
          },
          "check_in_location": {
            "type": "object",
            "description": "Check in location"
          },
          "check_out_location": {
            "type": "object",
            "description": "Check out location"
          },
          "status": {
            "type": "string",
            "description": "Status"
          },
          "device_id": {
            "type": "string",
            "description": "Device id"
          },
          "idempotency_key": {
            "type": "string",
            "description": "Idempotency key"
          },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "description": "Created at"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string",
            "description": "Updated at"
          }
        },
        "required": [
          "id",
          "salesman_id",
          "distributor_id",
          "check_in_at",
          "status",
          "created_at",
          "updated_at"
        ]
      },
...
    }
  }
}